Remove internal usage of std (#9)
* MSRV: bump from 1.63 to 1.81

* Remove internal usage of std

std was used for the Error trait and Duration type, both of which
are available in core.

* remove "std" feature.

the functionality "std" enabled in rustix-uring is now available without
the "std" feature. this feature also enabled the "std" feature for the
crates rustix and bitflags. downstream users of rustix-uring should
explicitly enable the "std" features for those crates if needed.

* release 0.4.0

breaking changes:

- MSRV 1.63 -> 1.81
- "std" feature removed